Move scroll checker into reusable hook
