+++ packages/core/components/Modal/BaseModal/index.tsx
@@ -3,8 +3,10 @@ import classNames from "classnames";
 import { noop } from "lodash";
 import * as React from "react";
 
-import styles from "./BaseModal.module.css";
 import { TertiaryButton } from "../../Buttons";
+import useCheckOverflowScroll from "../../../hooks/useCheckOverflowScroll";
+
+import styles from "./BaseModal.module.css";
 
 interface BaseModalProps {
     body: React.ReactNode;
@@ -21,32 +23,7 @@ interface BaseModalProps {
  */
 export default function BaseModal(props: BaseModalProps) {
     const { body, className, footer, title, onDismiss } = props;
-    const bodyContainerRef = React.useRef<HTMLDivElement>(null);
-    const [hasScroll, setHasScroll] = React.useState(false);
-
-    React.useEffect(() => {
-        const el = bodyContainerRef.current;
-        if (!el) return;
-
-        const checkScroll = () => {
-            const hasOverflowingContent = el.scrollHeight > el.clientHeight;
-            // we've reached the end of the content if the distance scrolled
-            // from the top (rounded up to account for sub-pixel differences)
-            // is equal to the total content height
-            const isAtBottomOfContent =
-                Math.ceil(el.scrollTop) + el.clientHeight >= el.scrollHeight;
-            setHasScroll(hasOverflowingContent && !isAtBottomOfContent);
-        };
-        checkScroll();
-
-        const observer = new ResizeObserver(checkScroll);
-        observer.observe(el);
-        el.addEventListener("scroll", checkScroll);
-        return () => {
-            observer.disconnect();
-            el.removeEventListener("scroll", checkScroll);
-        };
-    }, [body]);
+    const [scrollRef, hasScroll] = useCheckOverflowScroll<HTMLDivElement>([body]);
 
     const titleId = "base-modal-title";
     return (
@@ -67,7 +44,7 @@ export default function BaseModal(props: BaseModalProps) {
                 <TertiaryButton iconName="Cancel" onClick={onDismiss} title="" />
             </div>
             <div className={styles.bodyWrapper}>
-                <div className={styles.scrollableBody} ref={bodyContainerRef}>
+                <div className={styles.scrollableBody} ref={scrollRef}>
                     {body}
                 </div>
                 {hasScroll && <div className={styles.verticalGradient} />}
